+++ operator/pkg/controllers/cluster_controller.go
@@ -971,6 +971,14 @@ func (r *AIStoreReconciler) checkAISClusterReady(ctx context.Context, ais *aisv1
 		return
 	}
 
+	if ais.Status.ClusterID == "" {
+		ais.Status.ClusterID = smap.UUID
+		if err = r.patchStatus(ctx, ais); err != nil {
+			logger.Error(err, "Failed to update cluster ID")
+			return
+		}
+	}
+
 	return true, nil
 }